+++ cmd/mtping/kodata/LICENSE
