+++ cmd/mtping/kodata/VENDOR-LICENSE
