+++ cmd/ping/kodata/LICENSE
