+++ cmd/ping/kodata/VENDOR-LICENSE
