+++ cmd/ping/main.go
