--- config/200-jobrunner-clusterrole.yaml
@@ -1,1 +1,0 @@
-core/roles/jobrunner-clusterrole.yaml
