+++ config/200-mt-pingsource-clusterrole.yaml
@@ -1,0 +1,1 @@
+core/roles/mt-pingsource-clusterrole.yaml
