--- config/201-jobrunner-serviceaccount.yaml
@@ -1,1 +1,0 @@
-core/200-jobrunner-serviceaccount.yaml
