+++ config/201-mt-pingsource-serviceaccount.yaml
@@ -1,0 +1,1 @@
+core/200-mt-pingsource-serviceaccount.yaml
