+++ config/core/200-mt-pingsource-serviceaccount.yaml
@@ -1,4 +1,4 @@
-# Copyright 2019 The Knative Authors
+# Copyright 2020 The Knative Authors
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,41 +12,26 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRole
+apiVersion: v1
+kind: ServiceAccount
 metadata:
-  name: knative-eventing-jobrunner
+  name: eventing-mt-pingsource
+  namespace: knative-eventing
   labels:
     eventing.knative.dev/release: devel
-rules:
-  - apiGroups:
-      - ""
-    resources:
-      - "configmaps"
-    verbs:
-      - "get"
-      - "list"
-      - "watch"
-  - apiGroups:
-      - sources.knative.dev
-    resources:
-      - pingsources
-      - pingsources/status
-    verbs:
-      - get
-      - list
-      - watch
-      - patch
-  - apiGroups:
-      - sources.knative.dev
-    resources:
-      - pingsources/finalizers
-    verbs:
-      - "patch"
-  - apiGroups:
-      - ""
-    resources:
-      - events
-    verbs:
-      - "create"
-      - "patch"
+
+---
+apiVersion: rbac.authorization.k8s.io/v1
+kind: ClusterRoleBinding
+metadata:
+  name: eventing-mt-pingsource
+  labels:
+    eventing.knative.dev/release: devel
+subjects:
+  - kind: ServiceAccount
+    name: eventing-mt-pingsource
+    namespace: knative-eventing
+roleRef:
+  kind: ClusterRole
+  name: knative-eventing-mt-pingsource
+  apiGroup: rbac.authorization.k8s.io
