+++ config/core/deployments/controller.yaml
@@ -55,9 +55,9 @@ spec:
             value: knative.dev/eventing
           # PingSource
           - name: PING_IMAGE
-            value: ko://knative.dev/eventing/cmd/ping/adapter
-          - name: JOB_RUNNER_IMAGE
-            value: ko://knative.dev/eventing/cmd/ping/jobrunner
+            value: ko://knative.dev/eventing/cmd/ping
+          - name: MT_PING_IMAGE
+            value: ko://knative.dev/eventing/cmd/mt-ping
           # APIServerSource
           - name: APISERVER_RA_IMAGE
             value: ko://knative.dev/eventing/cmd/apiserver_receive_adapter
