+++ pkg/reconciler/pingsource/resources/receive_adapter.go
